% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_country.R
\docType{data}
\name{life_country}
\alias{life_country}
\title{Data for timelineGroup function in timelineS package}
\format{dataframe of name(character), country(character), gender(character), phase(character) and dates(date)}
\description{
Dates of birth and death, gender and phases
}
\keyword{timelineGroup}

