% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_exp.R
\docType{data}
\name{life_exp}
\alias{life_exp}
\title{Data for functions in timelineR package}
\format{dataframe of name(character), country(character), gender(character), and dates(date)}
\description{
Dates of birth and death, country, gender and names
}
\keyword{durCalc}
\keyword{durPlot,}
\keyword{durSummary,}

