% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mj_life.R
\docType{data}
\name{mj_life}
\alias{mj_life}
\title{Data for timelineS function in timelineS package}
\format{dataframe of events(character) and dates(date)}
\description{
Events and dates of Michael Jackson's life
}

